SP: adding helper function for validating board id
