+++ app/helper_function.py
@@ -1,0 +1,20 @@
+from flask import request, make_response, abort
+# helper function for validate key of input
+def validate_input_key_for_post_or_update():
+    request_board = request.get_json()
+    if "title" not in request_board or "owner" not in request_board:
+        abort(make_response({"details": "Invalid data"}, 400))
+    return request_board
+
+# helper function for validate id
+def get_board_or_abort(board_id):
+    try:
+        board_id = int(board_id)
+    except ValueError:
+        abort(make_response({"message": f"The task id {board_id} is invalid. The id must be integer."}, 400))
+    
+    boards = Board.query.all()
+    for board in boards:
+        if board.id == board_id:
+            return board
+    abort(make_response({"message": f"The task id {board_id} is not found"}, 404))
